Update PHP and Illuminate/Support versions
Bump the PHP requirement to allow both versions 7.2 and 8.*, as well as upgrade the Illuminate/Support package to versions 9.0 and 10.*. This ensures compatibility with a wider range of PHP versions and benefits from the latest features and improvements in Illuminate/Support. The new requirements provide greater flexibility and support for future updates. Fix any potential compatibility issues with GuzzleHttp by upgrading it to version 7.0. Also, update the ext-json dependency to allow any version. These changes enhance the overall stability and maintainability of the project.

Issue: N/A